--- common/clients/__init__.py
@@ -1,3 +1,0 @@
-from common.config import load_config
-
-config = load_config()
