+++ common/utility.py
@@ -184,3 +184,5 @@ class System:
 
         return (round(total_size / (1024 ** 3), 2), 'GB') if total_size > 1024 ** 3\
             else (round(total_size / (1024 ** 2), 2), 'MB')
+
+
