+++ unit3dup/media_manager/VideoManager.py
@@ -2,11 +2,12 @@
 import argparse
 import os
 
-from unit3dup.media_manager.models.qbitt import QBittorrent
-from unit3dup.media_manager.utility import UserContent
 from common.external_services.theMovieDB.core.api import DbOnline
-from unit3dup.upload import UploadBot
+from unit3dup.media_manager.common import UserContent
+
+from unit3dup.qbittorrent import QBittorrent
 from unit3dup.contents import Contents
+from unit3dup.upload import UploadBot
 from unit3dup.pvtVideo import Video
 from unit3dup import config
