+++ unit3dup/pvtVideo.py
@@ -3,9 +3,10 @@ import diskcache
 
 from common.external_services.imageHost import ImgBB, Freeimage, LensDump, ImageUploaderFallback
 from common.mediainfo import MediaFile
-from unit3dup import config
 from common.frames import VideoFrame
 from common.custom_console import custom_console
+
+from unit3dup import config
 
 offline_uploaders = []
